Improved logging and added control via options.
Using configargparse to get options from either files or the command
line to control both the logging level and other aspects of the code,
including where the datastore is and the name of the database.

Improved the logging to allow separate levels for for the console and
the log file.