add form to manage user information.
This change required the creation of another form which was very similar to the
create user form, but with some modifications for validation.
For example, if an admin is updating user information, (such as adding
to a group), they may not know or want to change the user password, and
would leave this field blank.
In order to do this, the form which was previously used to create a user
was broken up into smaller parts to allow modularization and re-use in
other forms. On the update user form, we validate the form after submitting,
using wtforms validate in the POST method rather than flask-wtforms validate_on_submit.
This allows us to add some validation functions to the field after verifying
some things on the Python side. For example, if we are updating a user's
password and their password is unique, the usual check for username uniqueness
would fail, so we want to exclude this if it is an update.