create job import form
This commit creates a form for importing jobs to the seamm.db and dashboard. The user should upload their job_data.json to complete the job import. This method assumes the job is in the original directory where the job was run. This commit also adds another css class and necessary stylings for form links from the navigation sidebar