add waitress server; front end improvements
this commit switches the time out method from taking place mostly on the front end (javascript) to taking place on the backend (python). Behavior is added so that when the front end encounters an expired token, the user is automatically logged out. The token is also refreshed automatically if the user is active within 30 minutes of the token expiring. The time out of the token is also set in the app configuration now, meaning this could be added as a configuration option on the command line